refactor(types): move ShareLimitAction enum to src/types/vuetorrent
