+++ src/constants/qbit/AppPreferences.ts
@@ -33,15 +33,6 @@ export enum FileLogAgeType {
   MONTHS = 1,
   YEARS = 2,
 }
-
-export enum ShareLimitAction {
-  DEFAULT = -1,
-  STOP_TORRENT = 0,
-  REMOVE_TORRENT = 1,
-  ENABLE_SUPERSEEDING = 2,
-  REMOVE_TORRENT_AND_FILES = 3,
-}
-
 export enum ProxyType {
   NONE = 'None',
   SOCKS4 = 'SOCKS4',
